+++ nerfstudio/process_data/base_converter_to_nerfstudio_dataset.py
@@ -1,0 +1,46 @@
+# Copyright 2022 The Nerfstudio Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Base class to process images or video into a nerfstudio dataset
+"""
+
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from pathlib import Path
+
+
+@dataclass
+class BaseConverterToNerfstudioDataset(ABC):
+    """Base class to process images or video into a nerfstudio dataset"""
+
+    data: Path
+    """Path the data, either a video file or a directory of images."""
+    output_dir: Path
+    """Path to the output directory."""
+    verbose: bool = False
+    """If True, print extra logging."""
+
+    def __post_init__(self) -> None:
+        self.output_dir.mkdir(parents=True, exist_ok=True)
+        self.image_dir.mkdir(parents=True, exist_ok=True)
+
+    @property
+    def image_dir(self) -> Path:  # pylint: disable=missing-function-docstring
+        return self.output_dir / "images"
+
+    @abstractmethod
+    def main(self) -> None:
+        """This method implements the conversion logic for each type of data"""
+        raise NotImplementedError("the main method for conversion needs to be implemented")
